Split group applications into own file, split members & applications from groups in swagger docs
